refactor: :recycle: outlined button をリファクタ
