+++ lib/core/common_widget/button/follow_or_unfollow_button.dart
@@ -5,7 +5,7 @@ import '../../../feature/user_profile/application/user_follow_service.dart';
 import '../../../feature/user_profile/application/user_profile_state.dart';
 import '../shimmer_widget.dart';
 import 'filled_button.dart';
-import 'primary_outlined_button.dart';
+import 'outlined_button.dart';
 
 class FollowOrUnfollowButton extends ConsumerWidget {
   const FollowOrUnfollowButton({
