--- lib/core/common_widget/button/primary_outlined_button.dart
@@ -1,40 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// primaryカラーを使用したoutlinedボタン
-class PrimaryOutlinedButton extends StatelessWidget {
-  const PrimaryOutlinedButton({
-    super.key,
-    required this.onPressed,
-    required this.text,
-  });
-
-  final VoidCallback? onPressed;
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextButton(
-      style: TextButton.styleFrom(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(48),
-        ),
-        side: BorderSide(
-          color: Theme.of(context).colorScheme.primary,
-        ),
-        backgroundColor: Theme.of(context).colorScheme.onPrimary,
-      ),
-      onPressed: onPressed,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: 24,
-        ),
-        child: Text(
-          text,
-          style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                color: Theme.of(context).colorScheme.primary,
-              ),
-        ),
-      ),
-    );
-  }
-}
