--- lib/core/common_widget/button/secondary_outlined_button.dart
@@ -1,40 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// 白黒を基調としたoutlinedボタン
-class SecondaryOutlinedButton extends StatelessWidget {
-  const SecondaryOutlinedButton({
-    super.key,
-    required this.onPressed,
-    required this.text,
-  });
-
-  final VoidCallback? onPressed;
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return OutlinedButton(
-      style: OutlinedButton.styleFrom(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(48).copyWith(),
-        ),
-        side: BorderSide(
-          color: Theme.of(context).colorScheme.tertiary,
-        ),
-        backgroundColor: Theme.of(context).colorScheme.onTertiary,
-      ),
-      onPressed: onPressed,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: 24,
-        ),
-        child: Text(
-          text,
-          style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                color: Theme.of(context).colorScheme.tertiary,
-              ),
-        ),
-      ),
-    );
-  }
-}
